+++ controllers/keda/scaledobject_controller_test.go
@@ -394,7 +394,8 @@ var _ = Describe("ScaledObjectController", func() {
 				err = k8sClient.Get(context.Background(), types.NamespacedName{Name: soName, Namespace: "default"}, so)
 				Expect(err).ToNot(HaveOccurred())
 				so.Status.HpaName = ""
-				return k8sClient.Status().Update(context.Background(), so)
+				so.Generation++ // since the reconcile ignores status update we force generation change to trigger reconcile
+				return k8sClient.Update(context.Background(), so)
 			}).ShouldNot(HaveOccurred())
 
 			// Wait until the hpaName is updated in the scaled object.
@@ -1179,7 +1180,7 @@ var _ = Describe("ScaledObjectController", func() {
 				return metav1.ConditionTrue
 			}
 			return so.Status.Conditions.GetPausedCondition().Status
-		}, 5*time.Second).Should(Or(Equal(metav1.ConditionFalse), Equal(metav1.ConditionUnknown)))
+		}, 5*time.Second).Should(Equal(metav1.ConditionFalse))
 
 		// set annotation
 		Eventually(func() error {
@@ -1204,7 +1205,7 @@ var _ = Describe("ScaledObjectController", func() {
 		Eventually(func() metav1.ConditionStatus {
 			err := k8sClient.Get(context.Background(), types.NamespacedName{Name: soName, Namespace: "default"}, so)
 			if err != nil {
-				return metav1.ConditionUnknown
+				return metav1.ConditionFalse
 			}
 			return so.Status.Conditions.GetPausedCondition().Status
 		}).WithTimeout(2 * time.Minute).WithPolling(10 * time.Second).Should(Equal(metav1.ConditionTrue))
@@ -1268,7 +1269,7 @@ var _ = Describe("ScaledObjectController", func() {
 				return metav1.ConditionTrue
 			}
 			return so.Status.Conditions.GetPausedCondition().Status
-		}, 5*time.Second).Should(Or(Equal(metav1.ConditionFalse), Equal(metav1.ConditionUnknown)))
+		}, 5*time.Second).Should(Equal(metav1.ConditionFalse))
 
 		// set annotation to true at first
 		Eventually(func() error {
@@ -1293,7 +1294,7 @@ var _ = Describe("ScaledObjectController", func() {
 		Eventually(func() metav1.ConditionStatus {
 			err := k8sClient.Get(context.Background(), types.NamespacedName{Name: soName, Namespace: "default"}, so)
 			if err != nil {
-				return metav1.ConditionUnknown
+				return metav1.ConditionFalse
 			}
 			return so.Status.Conditions.GetPausedCondition().Status
 		}).WithTimeout(2 * time.Minute).WithPolling(10 * time.Second).Should(Equal(metav1.ConditionTrue))
@@ -1322,7 +1323,7 @@ var _ = Describe("ScaledObjectController", func() {
 		Eventually(func() metav1.ConditionStatus {
 			err := k8sClient.Get(context.Background(), types.NamespacedName{Name: soName, Namespace: "default"}, so)
 			if err != nil {
-				return metav1.ConditionUnknown
+				return metav1.ConditionTrue
 			}
 			return so.Status.Conditions.GetPausedCondition().Status
 		}).WithTimeout(2 * time.Minute).WithPolling(10 * time.Second).Should(Equal(metav1.ConditionFalse))
@@ -1401,7 +1402,7 @@ var _ = Describe("ScaledObjectController", func() {
 		Eventually(func() metav1.ConditionStatus {
 			err := k8sClient.Get(context.Background(), types.NamespacedName{Name: soName, Namespace: "default"}, so)
 			if err != nil {
-				return metav1.ConditionUnknown
+				return metav1.ConditionFalse
 			}
 			return so.Status.Conditions.GetPausedCondition().Status
 		}).Should(Equal(metav1.ConditionTrue))
@@ -1410,7 +1411,7 @@ var _ = Describe("ScaledObjectController", func() {
 		Eventually(func() metav1.ConditionStatus {
 			err := k8sClient.Get(context.Background(), types.NamespacedName{Name: soName, Namespace: "default"}, so)
 			if err != nil {
-				return metav1.ConditionUnknown
+				return metav1.ConditionFalse
 			}
 			return so.Status.Conditions.GetPausedCondition().Status
 		}).WithTimeout(1 * time.Minute).WithPolling(10 * time.Second).Should(Equal(metav1.ConditionTrue))